+++ lib/amazonka/src/Amazonka/HTTP.hs
@@ -20,144 +20,149 @@ where
 import Amazonka.Core.Lens.Internal (to, (^?), _Just)
 import Amazonka.Data.Body (isStreaming)
 import Amazonka.Env hiding (auth)
-import Amazonka.Logger
+import Amazonka.Env.Hooks (Finality (..))
+import qualified Amazonka.Env.Hooks as Hooks
 import Amazonka.Prelude
 import Amazonka.Types
 import Amazonka.Waiter
 import Control.Exception as Exception
 import Control.Monad.Trans.Resource (liftResourceT, transResourceT)
 import qualified Control.Retry as Retry
-import qualified Data.List as List
+import Data.Foldable (traverse_)
 import qualified Data.Time as Time
+import Data.Typeable (Typeable)
 import qualified Network.HTTP.Conduit as Client.Conduit
 
 retryRequest ::
   forall m a withAuth.
   ( MonadResource m,
     AWSRequest a,
+    Typeable a,
+    Typeable (AWSResponse a),
     Foldable withAuth
   ) =>
   Env' withAuth ->
   a ->
   m (Either Error (ClientResponse (AWSResponse a)))
-retryRequest env x =
-  Retry.retrying policy shouldRetry attempt
-  where
-    rq = configureRequest env x
-    attempt _ = httpRequest env rq
+retryRequest env@Env {hooks} rq = do
+  rq' <- liftIO $ Hooks.request hooks env rq
+  cfgRq <- configureRequest env rq'
 
-    policy = retryStream rq <> retryService (service rq)
+  let attempt _ = httpRequest env cfgRq
+      policy = retryStream cfgRq <> retryService (service cfgRq)
+      Request
+        { service = Service {retry = Exponential {check = serviceRetryCheck}}
+        } = cfgRq
 
-    shouldRetry :: Retry.RetryStatus -> Either Error b -> m Bool
-    shouldRetry s = \case
-      Left r
-        | Just True <- r ^? transportErr -> writeLog "http_error" s >> return True
-        | Just m <- r ^? serviceErr -> writeLog m s >> return True
-      _other -> return False
-      where
-        transportErr =
-          _TransportError . to (retryCheck env (Retry.rsIterNumber s))
+      shouldRetry :: Retry.RetryStatus -> Either Error b -> m Bool
+      shouldRetry s =
+        liftIO . \case
+          Left r
+            | Just True <- r ^? transportErr ->
+                True <$ Hooks.requestRetry hooks env (cfgRq, "http_error", s)
+            | Just name <- r ^? serviceErr ->
+                True <$ Hooks.requestRetry hooks env (cfgRq, name, s)
+          _other -> pure False
+        where
+          transportErr =
+            _TransportError . to (retryCheck env (Retry.rsIterNumber s))
 
-        serviceErr =
-          _ServiceError . to serviceRetryCheck . _Just
+          serviceErr =
+            _ServiceError . to serviceRetryCheck . _Just
 
-        Request
-          { service = Service {retry = Exponential {check = serviceRetryCheck}}
-          } = rq
-
-    writeLog m s =
-      logDebug (logger env)
-        . mconcat
-        . List.intersperse " "
-        $ [ "[Retry " <> build m <> "]",
-            "after",
-            build (Retry.rsIterNumber s + 1),
-            "attempts."
-          ]
+  Retry.retrying policy shouldRetry attempt >>= \case
+    Left e -> Left e <$ liftIO (Hooks.error hooks env (Final, cfgRq, e))
+    Right a -> pure $ Right a
 
 awaitRequest ::
   ( MonadResource m,
     AWSRequest a,
+    Typeable a,
     Foldable withAuth
   ) =>
   Env' withAuth ->
   Wait a ->
   a ->
   m (Either Error Accept)
-awaitRequest env@Env {..} w@Wait {..} x = do
-  let rq = configureRequest env x
-      attempt _ = handleResult rq <$> httpRequest env rq
+awaitRequest env@Env {hooks} w rq = do
+  rq' <- liftIO $ Hooks.request hooks env rq
+  cfgRq <- configureRequest env rq'
+  w'@Wait {..} <- liftIO $ Hooks.wait hooks env w
 
-  Retry.retrying policy (check logger) attempt <&> \case
-    (AcceptSuccess, _) -> Right AcceptSuccess
-    (_, Left e) -> Left e
-    (a, _) -> Right a
-  where
-    policy =
-      Retry.limitRetries attempts
-        <> Retry.constantDelay (toMicroseconds delay)
+  let handleResult res = (fromMaybe AcceptRetry $ accept w' cfgRq res, res)
+      attempt _ = handleResult <$> httpRequest env cfgRq
+      policy =
+        Retry.limitRetries attempts
+          <> Retry.constantDelay (toMicroseconds delay)
 
-    check e n (a, _) = writeLog e n a >> return (retry a)
-      where
-        retry AcceptSuccess = False
-        retry AcceptFailure = False
-        retry AcceptRetry = True
+      check retryStatus (a, _) = do
+        liftIO $ Hooks.awaitRetry hooks env (cfgRq, w', a, retryStatus)
+        pure $ case a of
+          AcceptSuccess -> False
+          AcceptFailure -> False
+          AcceptRetry -> True
 
-    handleResult rq =
-      first (fromMaybe AcceptRetry . accept w rq)
-        . join (,)
+  Retry.retrying policy check attempt >>= \case
+    (AcceptSuccess, _) -> pure $ Right AcceptSuccess
+    (_, Left e) -> Left e <$ liftIO (Hooks.error hooks env (Final, cfgRq, e))
+    (a, _) -> pure $ Right a
 
-    writeLog l s a =
-      logDebug l
-        . mconcat
-        . List.intersperse " "
-        $ [ "[Await " <> build name <> "]",
-            build a,
-            "after",
-            build (Retry.rsIterNumber s + 1),
-            "attempts."
-          ]
-
--- | The 'Service' is configured + unwrapped at this point.
+-- | Make a one-shot request to AWS, using a configured 'Request'
+-- (which contains the 'Service', plus any overrides).
 httpRequest ::
   ( MonadResource m,
     AWSRequest a,
+    Typeable a,
     Foldable withAuth
   ) =>
   Env' withAuth ->
   Request a ->
   m (Either Error (ClientResponse (AWSResponse a)))
-httpRequest env@Env {..} x =
+httpRequest env@Env {hooks, manager, region} cfgRq =
   liftResourceT (transResourceT (`Exception.catches` handlers) go)
   where
     go = do
       time <- liftIO Time.getCurrentTime
 
-      rq <- case authMaybe env of
-        Nothing -> pure $! requestUnsigned x region
-        Just auth -> withAuth auth $ \a -> do
-          let Signed meta rq = requestSign x a region time
-          logTrace logger meta -- trace:Signing:Meta
-          pure $! rq
+      clientRq :: ClientRequest <-
+        liftIO . Hooks.clientRequest hooks env =<< case authMaybe env of
+          Nothing -> pure $! requestUnsigned cfgRq region
+          Just auth -> withAuth auth $ \a -> do
+            let s@(Signed _ rq) = requestSign cfgRq a region time
+            liftIO $ Hooks.signedRequest hooks env s
+            pure $! rq
 
-      logDebug logger rq -- debug:ClientRequest
-      rs <- Client.Conduit.http rq manager
+      rs <- Client.Conduit.http clientRq manager
+      liftIO $ Hooks.clientResponse hooks env (cfgRq, () <$ rs)
+      parsedRs <-
+        response
+          (Hooks.rawResponseBody hooks env)
+          (service cfgRq)
+          (proxy cfgRq)
+          rs
+      traverse_ (liftIO . Hooks.response hooks env . (cfgRq,)) parsedRs
+      pure parsedRs
 
-      logDebug logger rs -- debug:ClientResponse
-      response logger (service x) (proxy x) rs
-
+    handlers :: [Handler (Either Error b)]
     handlers =
       [ Handler $ err,
         Handler $ err . TransportError
       ]
       where
-        err e = logError logger e >> return (Left e)
+        err e = Left e <$ Hooks.error hooks env (NotFinal, cfgRq, e)
 
     proxy :: Request a -> Proxy a
     proxy _ = Proxy
 
-configureRequest :: AWSRequest a => Env' withAuth -> a -> Request a
-configureRequest Env {overrides} = request overrides
+-- Configures an AWS request `a` into its `Request a` form, applying
+-- service overrides from `env` and running hooks on the configured
+-- (Request a).
+configureRequest ::
+  (AWSRequest a, Typeable a, MonadIO m) => Env' withAuth -> a -> m (Request a)
+configureRequest env@Env {overrides, hooks} =
+  liftIO
+    . Hooks.configuredRequest hooks env
+    . request overrides
 
 retryStream :: Request a -> Retry.RetryPolicy
 retryStream Request {body} =